+++ src/services/auth/sponsor/sponsor-router.ts
@@ -1,77 +1,32 @@
 import { Router } from "express";
 import { Database } from "../../../database";
-import RoleChecker from "../../../middleware/role-checker";
-import { Role } from "../auth-models";
 import { StatusCodes } from "http-status-codes";
 import { sendEmail } from "../../ses/ses-utils";
 import jsonwebtoken from "jsonwebtoken";
 import { Config } from "../../../config";
+import { createSixDigitCode, encryptSixDigitCode} from "./sponsor-utils";
 import * as bcrypt from "bcrypt";
+import {AuthSponsorLoginValidator, AuthSponsorVerifyValidator} from "./sponsor-schema";
+
 const sponsorRouter = Router();
 
-// Get favorite events for an attendee
-sponsorRouter.get(
-    "/",
-    RoleChecker([Role.Enum.CORPORATE]),
-    async (req, res, next) => {
-        try {
-            const resumeUsers = await Database.REGISTRATION.find(
-                { hasResume: true },
-                { userId: 1 }
-            );
-            if (!resumeUsers) {
-                return res
-                    .status(StatusCodes.NOT_FOUND)
-                    .json({ error: "UserNotFound" });
-            }
-            return res.status(StatusCodes.OK).json(resumeUsers);
-        } catch (error) {
-            next(error);
-        }
-    }
-);
-
-function createSixDigitCode() {
-    let result = "";
-    const chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-    for (let i = 0; i < 6; i++) {
-        result += chars.charAt(Math.floor(Math.random() * chars.length));
-    }
-    return result;
-}
-
-function encryptSixDigitCode(sixDigitCode: string): string {
-    console.log("SixDigit: ", sixDigitCode);
-    const saltRounds = 10;
-
-    try {
-        const hash = bcrypt.hashSync(sixDigitCode, saltRounds);
-        return hash;
-    } catch (err) {
-        console.error("Error encrypting the code:", err);
-        throw err;
-    }
-}
-
 sponsorRouter.post("/login", async (req, res, next) => {
-    const { email } = req.body;
     try {
+        const { email } = AuthSponsorLoginValidator.parse(req.body);
         const sixDigitCode = createSixDigitCode();
         const expTime = Math.floor(Date.now() / 1000) + 300;
         const hashedVerificationCode = encryptSixDigitCode(sixDigitCode);
         await Database.AUTH_CODES.findOneAndUpdate(
             { email },
             {
-                $set: {
-                    hashedVerificationCode: hashedVerificationCode,
-                    expTime: expTime,
-                },
+                hashedVerificationCode: hashedVerificationCode,
+                expTime: expTime,
             },
             { upsert: true }
         );
         await sendEmail(
             email,
-            "RP-Sponor Email Verification!",
+            "R|P Sponsor Email Verification!",
             `Here is your verification code: ${sixDigitCode}`
         );
         return res.sendStatus(StatusCodes.CREATED);
@@ -81,24 +36,23 @@ sponsorRouter.post("/login", async (req, res, next) => {
 });
 
 sponsorRouter.post("/verify", async (req, res, next) => {
-    const { email, sixDigitCode } = req.body;
     try {
-        const sponsorData = await Database.AUTH_CODES.findOne({ email });
+        const { email, sixDigitCode } = AuthSponsorVerifyValidator.parse(req.body);
+        const sponsorData = await Database.AUTH_CODES.findOneAndDelete({ email });
         if (!sponsorData) {
-            return res.status(401).json({ message: "No Access" });
+            return res.sendStatus(StatusCodes.UNAUTHORIZED);
         }
         const { hashedVerificationCode, expTime } = sponsorData;
         if (Math.floor(Date.now() / 1000) > expTime) {
-            return res.status(401).json({ message: "Code expired" });
+            return res.sendStatus(StatusCodes.GONE);
         }
         const match = await bcrypt.compareSync(
             sixDigitCode,
             hashedVerificationCode
         );
         if (!match) {
-            return res.status(401).json({ message: "Incorrect Code" });
+            return res.sendStatus(StatusCodes.BAD_REQUEST);
         }
-        await Database.AUTH_CODES.deleteOne({ email });
         const token = jsonwebtoken.sign(
             {
                 email,
@@ -106,7 +60,7 @@ sponsorRouter.post("/verify", async (req, res, next) => {
             },
             Config.JWT_SIGNING_SECRET,
             {
-                expiresIn: Config.JWT_EXPIRATION_TIME,
+                expiresIn: (Math.floor(Date.now() / 1000)) + Config.JWT_EXPIRATION_TIME
             }
         );
         res.json({ token });
